requirements.txt
beautifulsoup4
selenium
openpyxl
pytest
python-dotenv